+++ config/routes.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 RecordingStudioUser::Engine.routes.draw do
+  resources :otp_codes, only: :show
+
   resource :profile, only: %i[show edit update], path: RecordingStudioUser.config.profile_route_path do
     get "sign-in-methods", to: "sign_in_methods#show", as: :sign_in_methods
     resources :identities, only: %i[destroy], param: :provider